Remove dead code and dead parameters
