+++ core/src/main/java/org/elasticsearch/action/update/TransportUpdateAction.java
@@ -153,7 +153,7 @@ public class TransportUpdateAction extends TransportInstanceSingleOperationActio
             return clusterState.routingTable().index(request.concreteIndex()).shard(request.getShardId().getId()).primaryShardIt();
         }
         ShardIterator shardIterator = clusterService.operationRouting()
-                .indexShards(clusterState, request.concreteIndex(), request.type(), request.id(), request.routing());
+                .indexShards(clusterState, request.concreteIndex(), request.id(), request.routing());
         ShardRouting shard;
         while ((shard = shardIterator.nextOrNull()) != null) {
             if (shard.primary()) {
